Finish unit tests, add decodeBlock and transaction list
Change the previous decodes to take and return a transaction
instead of a block.

Add a decodeBlock that will decode all of the transactions
in a block instead of just the first one.

Use the decode block in the /chain/blocks and chain/blockList
REST calls so that they will have all of the transactions
decoded.

Add the /chain/transactionList REST endpoint

Add unit tests for decodeBlock and transaction list.